+++ src/mvvm/base/navigator.c
@@ -339,6 +339,16 @@ ret_t navigator_close(const char* name) {
   char target[MAX_PATH + 1];
   return_value_if_fail(name != NULL, RET_BAD_PARAMS);
 
+  tk_snprintf(target, sizeof(target) - 1, "%s?%s=%s&force=true", NAVIGATOR_REQ_CLOSE, NAVIGATOR_ARG_NAME,
+              name);
+
+  return navigator_to(target);
+}
+
+ret_t navigator_request_close(const char* name) {
+  char target[MAX_PATH + 1];
+  return_value_if_fail(name != NULL, RET_BAD_PARAMS);
+
   tk_snprintf(target, sizeof(target) - 1, "%s?%s=%s", NAVIGATOR_REQ_CLOSE, NAVIGATOR_ARG_NAME,
               name);